+++ tests/test_data_models.py
@@ -1,7 +1,6 @@
 """Tests for the `data_models` module."""
 
 import inspect
-import json
 from collections.abc import Generator
 from pathlib import Path
 
@@ -235,43 +234,6 @@ class TestBenchmarkResult:
     def test_from_dict(self, config: dict, expected: BenchmarkResult) -> None:
         """Test that `BenchmarkResult.from_dict` works as expected."""
         assert BenchmarkResult.from_dict(config) == expected
-
-    def test_append_to_results(
-        self, benchmark_result: BenchmarkResult, results_path: Path
-    ) -> None:
-        """Test that `BenchmarkResult.append_to_results` works as expected."""
-        results_path.unlink(missing_ok=True)
-        results_path.touch(exist_ok=True)
-
-        benchmark_result.append_to_results(results_path=results_path)
-        json_str = json.dumps(
-            dict(
-                dataset=benchmark_result.dataset,
-                task=benchmark_result.task,
-                languages=benchmark_result.languages,
-                model=benchmark_result.model,
-                results=benchmark_result.results,
-                num_model_parameters=benchmark_result.num_model_parameters,
-                max_sequence_length=benchmark_result.max_sequence_length,
-                vocabulary_size=benchmark_result.vocabulary_size,
-                merge=benchmark_result.merge,
-                generative=benchmark_result.generative,
-                generative_type=benchmark_result.generative_type,
-                few_shot=benchmark_result.few_shot,
-                validation_split=benchmark_result.validation_split,
-                euroeval_version=benchmark_result.euroeval_version,
-                transformers_version=benchmark_result.transformers_version,
-                torch_version=benchmark_result.torch_version,
-                vllm_version=benchmark_result.vllm_version,
-                xgrammar_version=benchmark_result.xgrammar_version,
-            )
-        )
-        assert results_path.read_text() == f"\n{json_str}"
-
-        benchmark_result.append_to_results(results_path=results_path)
-        assert results_path.read_text() == f"\n{json_str}\n{json_str}"
-
-        results_path.unlink(missing_ok=True)
 
 
 class TestBenchmarkParametersAreConsistent: